more golang oriented package names
Trying to follow guidance of http://blog.golang.org/package-names
and golang's built in packages.  Sometimes a package location can
be ambiguous, but I tried to take advantage of the fact that the
package name is part of the identifier and create packages that
had common prefix names (like datapoint for example).

This also removes the megajson engine, but keeps the json
decoding abstraction.  Nobody is currently using the megajson
engine.